requirements.txt
aiohttp>=3.9.0
boto3>=1.34.0
Pillow>=10.0.0
